Extract chat discussion UI
