+++ js/chat-discussion-ui.js
@@ -1,0 +1,104 @@
+// chat-discussion-ui.js - DOM controls for multi-persona discussion flow
+
+import { state } from './state.js';
+import { CHAT_PERSONALITIES } from './constants.js';
+import { escapeHTML } from './utils.js';
+import { getCustomPersonalities } from './chat-personalities.js';
+
+export function showDiscussContinuePrompt(personas, originalPersonality, { onPersist } = {}) {
+  const container = document.getElementById('chat-messages');
+  if (!container) return;
+  const existing = container.querySelector('.chat-discuss-continue');
+  if (existing) existing.remove();
+
+  const prompt = document.createElement('div');
+  prompt.className = 'chat-discuss-continue';
+  prompt.innerHTML = '<input type="text" class="chat-discuss-steer" id="chat-discuss-steer" autocomplete="off" placeholder="Steer the debate (optional)..." onkeydown="if(event.key===\'Enter\'){event.preventDefault();continueDiscussion()}">' +
+    '<div class="chat-discuss-continue-actions">' +
+    '<button class="chat-discuss-continue-btn" onclick="continueDiscussion()">Continue</button>' +
+    '<button class="chat-discuss-done-btn" onclick="endDiscussion()">Done</button>' +
+    '</div>';
+  container.appendChild(prompt);
+  container.scrollTop = container.scrollHeight;
+
+  const steerInput = prompt.querySelector('.chat-discuss-steer');
+  if (steerInput) steerInput.focus();
+
+  state._discussionPersonas = personas;
+  state._discussionOriginalPersonality = originalPersonality;
+  onPersist?.(personas, originalPersonality);
+}
+
+export function removeDiscussContinuePrompt() {
+  const container = document.getElementById('chat-messages');
+  if (!container) return;
+  const el = container.querySelector('.chat-discuss-continue');
+  if (el) el.remove();
+}
+
+export function removeDiscussPersonaPicker() {
+  const picker = document.querySelector('.discuss-persona-picker');
+  if (picker) picker.remove();
+}
+
+export function showDiscussPersonaPicker() {
+  const allPersonas = [
+    ...CHAT_PERSONALITIES.map(p => ({ id: p.id, name: p.name, icon: p.icon })),
+    ...getCustomPersonalities().map(p => ({ id: p.id, name: p.name, icon: p.icon || '\u270f\ufe0f' }))
+  ];
+  if (allPersonas.length < 2) return;
+
+  const existing = document.querySelector('.discuss-persona-picker');
+  if (existing) { existing.remove(); return; }
+
+  const container = document.querySelector('.chat-input-area');
+  if (!container) return;
+
+  const activePersonaIds = new Set();
+  for (const m of state.chatHistory) {
+    if (m.role === 'assistant' && m.personalityName) {
+      const bp = CHAT_PERSONALITIES.find(p => p.name === m.personalityName);
+      if (bp) activePersonaIds.add(bp.id);
+      else {
+        const cp = getCustomPersonalities().find(p => p.name === m.personalityName);
+        if (cp) activePersonaIds.add(cp.id);
+      }
+    }
+  }
+  const hasActive = activePersonaIds.size > 0;
+  const needsOne = hasActive && activePersonaIds.size < 2;
+
+  const picker = document.createElement('div');
+  picker.className = 'discuss-persona-picker';
+  picker.innerHTML = `
+    <div class="discuss-picker-header">${needsOne ? 'Add another persona to the debate' : 'Pick two personas to debate'}</div>
+    <div class="discuss-picker-list">
+      ${allPersonas.map(p => {
+        const isActive = activePersonaIds.has(p.id);
+        const checked = isActive ? ' checked' : '';
+        const locked = isActive && needsOne;
+        return `<label class="discuss-picker-item${locked ? ' locked' : ''}">
+        <input type="checkbox" value="${escapeHTML(p.id)}" data-name="${escapeHTML(p.name)}" data-icon="${escapeHTML(p.icon)}"${checked}${locked ? ' disabled' : ''} data-locked="${locked ? '1' : ''}">
+        <span>${p.icon} ${escapeHTML(p.name)}</span>
+      </label>`;
+      }).join('')}
+    </div>
+    <button class="discuss-picker-start"${needsOne ? '' : ' disabled'} onclick="startDiscussionFromPicker()">${needsOne ? 'Add to Discussion' : 'Start Debate'}</button>`;
+
+  function updatePickerState() {
+    const lockedCount = picker.querySelectorAll('input[data-locked="1"]').length;
+    const checkedCount = picker.querySelectorAll('input:checked:not([data-locked="1"])').length;
+    const total = lockedCount + checkedCount;
+    const startBtn = picker.querySelector('.discuss-picker-start');
+    startBtn.disabled = total !== 2;
+    if (total >= 2) {
+      picker.querySelectorAll('input:not(:checked):not([data-locked="1"])').forEach(cb => cb.disabled = true);
+    } else {
+      picker.querySelectorAll('input:not([data-locked="1"])').forEach(cb => cb.disabled = false);
+    }
+  }
+  picker.addEventListener('change', updatePickerState);
+  updatePickerState();
+
+  container.insertBefore(picker, container.firstChild);
+}
