+++ tests/test-correctness-phase2.js
@@ -136,6 +136,7 @@ const pwaAppShellAssets = [
   '/js/chat-panel.js',
   '/js/chat-discussion.js',
   '/js/chat-discussion-state.js',
+  '/js/chat-discussion-ui.js',
   '/js/chat-onboarding.js',
   '/js/chat-empty-state.js',
   '/js/chat-render.js',
